Correct behaviour for ICH, DCH, IL, DL.
Earlier behaviour was modeled after XTerm, but recent versions of
XTerm have corrected some of the deviations (for tests of IL, DL)
of Vttest screen output, so we move our target from XTerm
bug-compatibility to absolute Vttest correctness.

The cause of deviations was that the sequences were not always
discarded if received at a screen position outside the margins.