Update testbase.sh
This fixes the usage of ps here.  Not only does this make it work on FreeBSD, but it also makes the usage POSIX-compliant:
https://pubs.opengroup.org/onlinepubs/9699919799/utilities/ps.html

Please confirm that the behavior is not changed on GNU/Linux before merging. (I don't have any easy way to test on GNU/Linux)
Also, more patches may come in as I come across more things which prevent me from running the test scripts.