Tune shell program resolution to use $SHELL if no shell is configured
The previous behaviour used 'bash' as a hardwired default if there was
no shell given (either as a -shell argument, or a last argument after
all options, or configured as the Zutty.shell resource).

The new behaviour is that in this case, the value of the SHELL
environment variable is used.

In addition, the resolved shell program path is set as the value of
SHELL (seen by the spawned shell program), unless it is not listed as
a valid shell in /etc/shells. Previously, "invalid" shell programs
caused SHELL to be unset, but SHELL could be left set to a 'bare'
non-resolved value.