Resize: observe south-west gravity
Scroll content before/after resize so bottom-most lines are not lost
on decreasing height; also try to fill the bottom gap on increasing
height as long as there is history to scroll down.  For obvious
reasons, these need to be done on opposite ends (before/after) of the
frame resize itself (which maintains northwest gravity, i.e., the
topmost row would stay put).

On a height decrease, we assume that there is nothing valuable below
the cursor. This might be false if the screen has been painted by an
interactive program moving the cursor around (as opposed to just
command output). However, interactive programs will repaint their UI
on the terminal geometry change.