Remove redundant call to FT_Render_Glyph
The font character is loaded with FT_LOAD_RENDER right before this, so
an explicit render call should not be needed. Further, recent FreeType
(2.11.0) seems to return an error (FT_Err_Cannot_Render_Glyph) in this
case, which is probably a bug in FreeType, but is hereby worked around.

See #6 for further info.