Correct logic around VT100 xenl-quirk
Add guard for illegal condition of being in the "invisible" off-by-one
row before trying to insert a character. In such a case, unless we
have arrived there via a directly preceding previous character
placement and auto-wrap is on, the new character will be discarded.
Also add guard for running csi_ED and csi_EL with the cursor in this
off-screen column (we need posX + 1 to evaluate to no more than nCols).

This condition does not normally occur, but can be evoked by
pathological (but nevertheless real-world) sequences, such as cat'ing
random binary data into the terminal.

Adding the guards substitutes the relevant parts of the previous commit
37d3781, which are now reverted in exchange for a more relevant fix.

Another subtle change is that we drop our ENQ (0x05/Enquiry) response.
This is generally an anachronism, but more specifically, it causes
problematic behaviour in the same real-world case of cat'ing large
binary files. Since the binary will inevitably contain ENQ bytes with
some random frequency, Zutty would emit its response that nobody will
drain. This output eventually fills up the outbound pipe, blocking
Zutty in a writePty().  The behaviour is technically correct, but the
user experience (Zutty stopped responding) is "highly unfortunate".

Closes #47
Closes #48
Closes #51