Correct distinction between 'raw' and 'snapped' selection rect
The code was slightly unclear on whether the selection rect contains
the 'raw' coords (closely reflecting the mouse actions) or the bounds
after snap-to word/line has taken place. This led to faulty behaviour
as exemplified by #55, due to having used the snapped coords where the
raw coords would have been correct.

Fix this by making Frame own the snapTo setting, and by only storing
the 'raw' selection rect and treating the snapped coords as ephemeral
(computed on demand). In practice, the snap-to operation is performed
the same number of times, so this does not even impact performance.

Closes #55