Correct last-column/wrap-pending state subtleties
Extend test suite with scripts mentioned in #57, plus verify the
output of wraptest (https://github.com/mattiase/wraptest) in Zutty.
This output is (as of the present commit) perfectly identical to the
output of XTerm.

As a bonus, add support for ED (Erase in Display) with arg 3, which
is an XTerm extension to erase the whole display including the
scrollback buffer.

Closes #57