Be lenient with POLLHUP signalling EOF from pty while starting up
On certain non-Linux systems (e.g., OpenBSD), poll () sets POLLHUP in
.revents to signal EOF, and this can happen before the first byte of
data arrives from the pty slave (written by the shell). That is, the
pty fd appears to have been closed from the other end, but this is
only a temporary condition.

The current (lack of) handling this leads, on certain systems, to
sporadic "startup errors" of Zutty wherein the program's window
appears and immediately disappears. To avoid this, unify our handlers
of POLLIN and POLLHUP, and exit on an EOF (which is what we get on a
read () after a POLLHUP) only after a successful read has happened
before, else keep looping.

Also, for the sake of portability, honor EINTR and retry the poll ()
instead of giving up.