Remove -march and -mtune from compiler options
These are not universally supported across all platforms, operating
systems and compilers Zutty targets (e.g., OpenBSD with clang on arm64);
further, their usage results in a potentially non-transferable binary,
so packagers of Zutty would have to remove them in any case. It is safer
to not include them as a default; those attempting to perf-tune Zutty
should be savvy enough to add them back.