Do not echo invalid input on DECRQSS (Request Status String)
Echoing invalid input to the console allows a malicious console
application to invoke an arbitrary command. We now respond to such
invalid requests with a cancel, just like xterm.

Thanks to Carter Sande of duodecima.technology for calling my
attention to this issue.

See also CVE-2008-2383 and the discussion on Debian bug #510030.