Upgrade WAF to version 2.0.25
As before, we maintain a manual change of the shebang (first line)
calling python3 instead of python. This is so that the script will
still run on legacy systems without python3 symlinked to python.

From the Debian wiki [https://wiki.debian.org/Python]:

    "No packaged scripts should depend on the existence of
    '/usr/bin/python': if they do, that is a bug [...].
    You can use the 'python-is-python3' or 'python-is-python2'
    packages to restore an appropriate '/usr/bin/python'
    symlink for third-party or legacy scripts."

In other words, we do not want to depend on such a package, and we
require python3 to run the configure script.