Fix clearing glyph area before rendering overlay font glyph
Using bh and bw as bounds is not strictly correct, since that might be
smaller than the entire glyph area, so in theory part of the
overridden glyph could remain visible.

This is a theoretical correctness issue - in practice, the overlay
font will contain a bold/italic/boldItalic variant of the glyph
already there, so its geometry will not be smaller than the existing
glyph and thus it will most likely fully overwrite the pre-rendered
content.

Lifted from b945fbd on ttrei's fork - thanks!

(Also fix some code formatting.)